Move to a much stricter pyright configuration
This moves us from using the "standard" type-checking mode (which has a
lot of the diagnostic rules disabled, which meant we were almost in a
whitelist-based system, where most diagnostic rules were disabled unless
explicitly enabled) to "all" (where all diagnotic rules are enabled,
except for those explicitly disabled, moving us to a blacklist-based
system).

The current configuration here is very strict and might not fit all
code-bases, but it will work very well for new projects, as this
strictness, while slightly annoying sometimes, forces a lot better and
more correct typing to the previous set of rules.

This template still leaves the rules for reporting unknown types
disabled, as these are incredibly strict and while they can definitely
be beneficial, for most people it's too big of an annoyance for what
it's worth.